[CP-2397][Multidevice] empty Select device screen
